Move logic to open signing window into its own hook for reusage
